requirements.txt
requests==2.32.3
numpy==1.26.4
faiss-cpu==1.8.0
scikit-learn==1.5.2
tqdm==4.66.5